+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 
 - Option to listen on a socket instead of HOST:PORT.
 - Vision tool to change active tokens.
+- Vision lock button to initiative to only show vision of current actor.
 
 ## [0.15.1] - 2019-05-15